refactor(components): move filter logic into hook
